requirements.txt
tensorflow
tensorflow-keras
Pillow
rasterio
numpy
geopandas
shapely
mlflow
matplotlib
colorlog
opencv-python